+++ src/main/java/com/podio/rating/RatingAPI.java
@@ -38,7 +38,7 @@ public class RatingAPI extends BaseAPI {
 	 * 
 	 * @see RatingValue
 	 */
-	public int createRating(Reference reference, RatingType type, int value) {
+	public long createRating(Reference reference, RatingType type, int value) {
 		return getResourceFactory()
 				.getApiResource("/rating/" + reference.toURLFragment() + type)
 				.post(Entity.entity(Collections.singletonMap("value", value),
@@ -65,7 +65,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object that should be liked
 	 * @return The id of the newly created rating
 	 */
-	public int like(Reference reference) {
+	public long like(Reference reference) {
 		return createRating(reference, RatingType.LIKE, RatingValue.LIKE);
 	}
 
@@ -76,7 +76,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int yes(Reference reference) {
+	public long yes(Reference reference) {
 		return createRating(reference, RatingType.YESNO, RatingValue.YESNO_YES);
 	}
 
@@ -87,7 +87,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int no(Reference reference) {
+	public long no(Reference reference) {
 		return createRating(reference, RatingType.YESNO, RatingValue.YESNO_NO);
 	}
 
@@ -98,7 +98,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int approves(Reference reference) {
+	public long approves(Reference reference) {
 		return createRating(reference, RatingType.APPROVED,
 				RatingValue.APPROVED_APPROVES);
 	}
@@ -110,7 +110,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int disapproves(Reference reference) {
+	public long disapproves(Reference reference) {
 		return createRating(reference, RatingType.APPROVED,
 				RatingValue.APPROVED_DISAPPROVES);
 	}
@@ -124,7 +124,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The number of stars to give, between 1 and 5
 	 * @return The id of the newly created rating
 	 */
-	public int fivestar(Reference reference, int stars) {
+	public long fivestar(Reference reference, int stars) {
 		return createRating(reference, RatingType.FIVESTAR, stars);
 	}
 
@@ -135,7 +135,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int attend(Reference reference) {
+	public long attend(Reference reference) {
 		return createRating(reference, RatingType.RSVP, RatingValue.RSVP_ATTEND);
 	}
 
@@ -146,7 +146,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int maybeAttend(Reference reference) {
+	public long maybeAttend(Reference reference) {
 		return createRating(reference, RatingType.RSVP,
 				RatingValue.RSVP_MAYBE_ATTEND);
 	}
@@ -158,7 +158,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int notAttend(Reference reference) {
+	public long notAttend(Reference reference) {
 		return createRating(reference, RatingType.RSVP,
 				RatingValue.RSVP_NOT_ATTEND);
 	}
@@ -170,7 +170,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int thumbsUp(Reference reference) {
+	public long thumbsUp(Reference reference) {
 		return createRating(reference, RatingType.THUMBS, RatingValue.THUMBS_UP);
 	}
 
@@ -181,7 +181,7 @@ public class RatingAPI extends BaseAPI {
 	 *            The reference to the object
 	 * @return The id of the newly created rating
 	 */
-	public int thumbsDown(Reference reference) {
+	public long thumbsDown(Reference reference) {
 		return createRating(reference, RatingType.THUMBS,
 				RatingValue.THUMBS_DOWN);
 	}
