+++ src/main/java/com/podio/search/SearchResult.java
@@ -10,7 +10,7 @@ import org.joda.time.DateTime;
 
 public class SearchResult {
         
-	private Integer id;
+	private Long id;
 
 	private ReferenceTypeSearchInApp type;
 
@@ -48,11 +48,11 @@ public class SearchResult {
 	 */
 	private String link;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
